#REFACTOR #PROMPT #MODEL_OUTPUT
Prompt Category:
Refactoring the feature.

Prompt:
The sidebar is collapsing with the other design. Please add the menu button on the navbar from where we open or close the sidebar.

User Observation:
The model response generates the menu button on the navbar but open and close of the sidebar is still not working and all the design is collapsing.
Response ID:
54088e0c-a91e-45a5-bc64-cd72cb99437f