#REFACTOR #PROMPT #MODEL_OUTPUT
Prompt Category:
Refactoring the feature.

Prompt:
Great Now you  have to convert the Sidebar.jsx design into the Header.jsx you have to add all the links on the Header.jsx which are currently on the Sidebar.jsx component and also give the spacing between the Header.jsx and remining components from the top because data is hidden back side of the Header.jsx

User Observation:
The sidebar converted to the Navbar successfully but still have issue of spacing.

Response ID:
8f460d1e-2de9-49af-98df-f9fb74b1ad5b