#REQ_DEV #PROMPT #MODEL_OUTPUT
Prompt Category:
Developing the new feature.

Prompt:
Great on the Profile.jsx component add the login signup button and when the user clicks on the Signup You have to allow the user to signup and than move to the login page after succesfully login move the user to the Dashboard.jsx component.
User Observation:
The model throws the error on the screen which is due to not define the `logout`.

Response ID:
3688d165-caa9-49d5-8065-d43bcbf68971